+++ frontend/src/components/notifications/NotificationItem.tsx
@@ -1,0 +1,108 @@
+import { Link } from 'react-router-dom';
+import { useTranslation } from 'react-i18next';
+import { Trash2 } from 'lucide-react';
+import type { Notification } from '../../types/notification';
+import Avatar from '../common/Avatar';
+import { formatDistanceToNow } from '../../utils/timeFormat';
+
+function getNotificationLink(notification: Notification): string {
+  switch (notification.type) {
+    case 'post':
+    case 'like':
+    case 'comment':
+      return notification.post_id ? `/posts/${notification.post_id}` : '/';
+    case 'follow':
+      return `/profile/${notification.actor.username}`;
+    case 'message':
+      return '/chat';
+    case 'answer':
+      return notification.question_id ? `/qa/${notification.question_id}` : '/';
+    case 'badge':
+      return `/profile/${notification.actor.username}`;
+    default:
+      return '/';
+  }
+}
+
+function getNotificationMessage(notification: Notification, t: (key: string, opts?: Record<string, string>) => string): string {
+  switch (notification.type) {
+    case 'post':
+      return t('notifications.newPost', { name: notification.actor.name });
+    case 'message':
+      return t('notifications.newMessage', { name: notification.actor.name });
+    case 'like':
+      return t('notifications.newLike', { name: notification.actor.name });
+    case 'comment':
+      return t('notifications.newComment', { name: notification.actor.name });
+    case 'follow':
+      return t('notifications.newFollow', { name: notification.actor.name });
+    case 'answer':
+      return t('notifications.newAnswer', { name: notification.actor.name });
+    case 'badge':
+      return t('notifications.newBadge');
+    default:
+      return '';
+  }
+}
+
+interface NotificationItemProps {
+  notification: Notification;
+  onMarkAsRead: (id: number) => void;
+  onDelete: (id: number) => void;
+}
+
+export default function NotificationItem({ notification, onMarkAsRead, onDelete }: NotificationItemProps) {
+  const { t } = useTranslation();
+
+  return (
+    <li
+      className={`flex items-start gap-3 p-4 rounded-lg border transition-colors ${
+        !notification.read
+          ? 'bg-gray-800/50 border-gray-700'
+          : 'bg-gray-900 border-gray-800'
+      }`}
+    >
+      <Link
+        to={getNotificationLink(notification)}
+        onClick={() => {
+          if (!notification.read) onMarkAsRead(notification.id);
+        }}
+        className="flex items-start gap-3 flex-1 min-w-0"
+      >
+        <Avatar
+          name={notification.actor.name}
+          avatarUrl={notification.actor.avatar_url}
+          size="sm"
+        />
+        <div className="flex-1 min-w-0">
+          <p className="text-sm text-gray-100">
+            {getNotificationMessage(notification, t)}
+          </p>
+          {(notification.type === 'post' || notification.type === 'like' || notification.type === 'comment') && notification.post && (
+            <p className="text-xs text-gray-400 truncate mt-0.5">
+              {notification.post.title}
+            </p>
+          )}
+          {notification.type === 'answer' && notification.question && (
+            <p className="text-xs text-gray-400 truncate mt-0.5">
+              {notification.question.title}
+            </p>
+          )}
+          <p className="text-xs text-gray-500 mt-1">
+            {formatDistanceToNow(notification.created_at)}
+          </p>
+        </div>
+        {!notification.read && (
+          <span className="w-2 h-2 bg-blue-500 rounded-full mt-2 shrink-0" aria-hidden="true" />
+        )}
+      </Link>
+      <button
+        onClick={() => onDelete(notification.id)}
+        className="p-1.5 text-gray-500 hover:text-red-400 transition-colors rounded-md shrink-0"
+        aria-label={t('notifications.deleteNotification')}
+      >
+        <Trash2 className="w-4 h-4" aria-hidden="true" />
+      </button>
+    </li>
+  );
+}
